Moved ModelTrainer and ModelTrainingArguments into the model folder.
